Fix to make the code build on FreeBSD/alpha 4.9.
git-svn-id: https://gxemul.svn.sourceforge.net/svnroot/gxemul/gxemul/trunk@5397 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/main/StateVariable.cc
@@ -379,7 +379,7 @@ string StateVariable::EvaluateExpression(const string& expression,
 	
 	// Remove leading and trailing spaces:
 	while (result.size() > 0 && result[0] == ' ')
-		result.erase(0);
+		result.erase((size_t) 0);
 	while (result.size() > 0 && result[result.size() - 1] == ' ')
 		result.erase(result.size()-1);
